Added GVRScriptBehavior scripting component
Allows a script to be attached as a component. Sets up event listeners
for picking and draw frame events and directs them to the script.
Directs all events on the scene object to the script.

Modified script manager to attach GVRScriptBehavior component when a
script bundle is loaded. Added onAttach, onDetach script events.