Fix ColliderGroup picking (#1874)
* Fix GVRColliderGroup picking

ColliderGroup must pass its owner object to the child colliders (which may not have owners). In this case it should use the mesh of the ColliderGroup owner.

* Fix ColliderGroup picking, getPickListChanged()

ColliderGroup must pass its owner object to the child colliders (which may not have owners). In this case it should use the mesh of the ColliderGroup owner.

I also added getPickListChanged() to determine whether the pick list has changed since last frame. I tested with 3D cursor sample and it works OK.

* Add GVRComponent.IComponentGroup

* Remove 3DCursor SelectableGroup